+++ influx-writer/influx-writer.py
@@ -6,7 +6,7 @@ import os
 from datetime import datetime
 sys.path.append(os.path.join(os.path.dirname(os.path.realpath(__file__)), ".."))
 from redis import Redis
-from redis_pubsub import RedisPubSubSubscriber, RedisPubSubPublisher, NoSubscriberNameSetException, WriteFailException, NoSubscribersException
+from listenstore.redis_pubsub import RedisPubSubSubscriber, RedisPubSubPublisher, NoSubscriberNameSetException, WriteFailException, NoSubscribersException
 from influxdb import InfluxDBClient
 from influxdb.exceptions import InfluxDBClientError, InfluxDBServerError
 import ujson
@@ -14,7 +14,7 @@ import logging
 from listen import Listen
 from time import time, sleep
 import config
-from listenstore.listenstore import InfluxListenStore
+from listenstore import InfluxListenStore
 
 REPORT_FREQUENCY = 5000
 SUBSCRIBER_NAME = "in"
