+++ redis-consumer/redis-consumer.py
@@ -4,13 +4,13 @@ import sys
 import os
 sys.path.append(os.path.join(os.path.dirname(os.path.realpath(__file__)), ".."))
 from redis import Redis
-from redis_pubsub import RedisPubSubSubscriber, NoSubscriberNameSetException, WriteFailException
+from listenstore.redis_pubsub import RedisPubSubSubscriber, NoSubscriberNameSetException, WriteFailException
 import config
 
 import ujson
 import logging
 from listen import Listen
-from listenstore.listenstore import PostgresListenStore
+from listenstore import PostgresListenStore
 from time import time, sleep
 
 REPORT_FREQUENCY = 5000
