+++ webserver/views/index.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 from flask import Blueprint, render_template, current_app, redirect, url_for
 from flask_login import current_user
 from webserver.redis_connection import _redis
-from redis_pubsub import RedisPubSubPublisher
+from listenstore.redis_pubsub import RedisPubSubPublisher
 import os
 import subprocess
 import locale
